+++ modules/terracotta-core/src/test/kotlin/io/github/beduality/terracotta/core/model/metadata/detector/ProjectMetadataLoaderTest.kt
@@ -1,4 +1,4 @@
-package io.github.beduality.terracotta.core.detect
+package io.github.beduality.terracotta.core.model.metadata.detector
 
 import io.github.beduality.terracotta.core.model.TerracottaEnvironment
 import io.github.beduality.terracotta.core.model.metadata.ProjectMetadataSource