+++ modules/terracotta-core/src/test/kotlin/io/github/beduality/terracotta/core/model/metadata/detector/adapters/ReadmeMetadataDetectorTest.kt
@@ -1,4 +1,4 @@
-package io.github.beduality.terracotta.core.detect.adapters
+package io.github.beduality.terracotta.core.model.metadata.detector.adapters
 
 import io.github.beduality.terracotta.core.model.metadata.ProjectMetadataContext
 import io.github.beduality.terracotta.core.model.metadata.ProjectMetadataSource