+++ modules/terracotta-gradle-plugin/src/main/kotlin/io/github/beduality/terracotta/gradle/detector/GradleGameVersionMetadataDetector.kt
@@ -1,7 +1,7 @@
-package io.github.beduality.terracotta.gradle.detect
+package io.github.beduality.terracotta.gradle.detector
 
-import io.github.beduality.terracotta.core.detect.ProjectMetadataDetector
-import io.github.beduality.terracotta.core.detect.adapters.GameVersionNormalizer
+import io.github.beduality.terracotta.core.model.metadata.detector.ProjectMetadataDetector
+import io.github.beduality.terracotta.core.model.metadata.detector.adapters.GameVersionNormalizer
 import io.github.beduality.terracotta.core.model.metadata.ProjectMetadataContext
 import io.github.beduality.terracotta.core.model.metadata.TerracottaProjectMetadata
 import io.github.beduality.terracotta.core.model.projectfile.ProjectFileCache