+++ modules/terracotta-gradle-plugin/src/test/kotlin/io/github/beduality/terracotta/gradle/detector/GradleGameVersionMetadataDetectorTest.kt
@@ -1,4 +1,4 @@
-package io.github.beduality.terracotta.gradle.detect
+package io.github.beduality.terracotta.gradle.detector
 
 import io.github.beduality.terracotta.core.model.metadata.ProjectMetadataContext
 import io.github.beduality.terracotta.core.model.metadata.ProjectMetadataSource